Removed legacy index.cache.json support
