+++ cmd/thanos/compact.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"os"
 	"path"
-	"path/filepath"
 	"strconv"
 	"strings"
 	"time"
@@ -16,7 +15,6 @@ import (
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/oklog/run"
-	"github.com/oklog/ulid"
 	"github.com/opentracing/opentracing-go"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
@@ -25,26 +23,18 @@ import (
 	"github.com/prometheus/common/route"
 	"github.com/prometheus/prometheus/tsdb"
 	"github.com/thanos-io/thanos/pkg/block"
-	"github.com/thanos-io/thanos/pkg/block/indexheader"
-	"github.com/thanos-io/thanos/pkg/block/metadata"
 	"github.com/thanos-io/thanos/pkg/compact"
 	"github.com/thanos-io/thanos/pkg/compact/downsample"
 	"github.com/thanos-io/thanos/pkg/component"
 	"github.com/thanos-io/thanos/pkg/extflag"
 	"github.com/thanos-io/thanos/pkg/extprom"
 	extpromhttp "github.com/thanos-io/thanos/pkg/extprom/http"
-	"github.com/thanos-io/thanos/pkg/objstore"
 	"github.com/thanos-io/thanos/pkg/objstore/client"
 	"github.com/thanos-io/thanos/pkg/prober"
 	"github.com/thanos-io/thanos/pkg/runutil"
 	httpserver "github.com/thanos-io/thanos/pkg/server/http"
 	"github.com/thanos-io/thanos/pkg/ui"
 	"gopkg.in/alecthomas/kingpin.v2"
-)
-
-const (
-	metricIndexGenerateName = "thanos_compact_generated_index_total"
-	metricIndexGenerateHelp = "Total number of generated indexes."
 )
 
 var (
@@ -265,7 +255,6 @@ func runCompact(
 	var (
 		compactDir      = path.Join(conf.dataDir, "compact")
 		downsamplingDir = path.Join(conf.dataDir, "downsample")
-		indexCacheDir   = path.Join(conf.dataDir, "index_cache")
 	)
 
 	if err := os.RemoveAll(downsamplingDir); err != nil {
@@ -344,17 +333,6 @@ func runCompact(
 
 	g.Add(func() error {
 		defer runutil.CloseWithLogOnErr(logger, bkt, "bucket client")
-
-		// Generate index files.
-		// TODO(bwplotka): Remove this in next release.
-		if conf.generateMissingIndexCacheFiles {
-			if err := sy.SyncMetas(ctx); err != nil {
-				return err
-			}
-			if err := genMissingIndexCacheFiles(ctx, logger, reg, bkt, sy.Metas(), indexCacheDir); err != nil {
-				return err
-			}
-		}
 
 		if !conf.wait {
 			return compactMainFn()
@@ -439,96 +417,6 @@ func runCompact(
 	return nil
 }
 
-// genMissingIndexCacheFiles scans over all blocks, generates missing index cache files and uploads them to object storage.
-func genMissingIndexCacheFiles(ctx context.Context, logger log.Logger, reg *prometheus.Registry, bkt objstore.Bucket, metas map[ulid.ULID]*metadata.Meta, dir string) error {
-	genIndex := promauto.With(reg).NewCounter(prometheus.CounterOpts{
-		Name: metricIndexGenerateName,
-		Help: metricIndexGenerateHelp,
-	})
-
-	if err := os.RemoveAll(dir); err != nil {
-		return errors.Wrap(err, "clean index cache directory")
-	}
-	if err := os.MkdirAll(dir, 0777); err != nil {
-		return errors.Wrap(err, "create dir")
-	}
-
-	defer func() {
-		if err := os.RemoveAll(dir); err != nil {
-			level.Error(logger).Log("msg", "failed to remove index cache directory", "path", dir, "err", err)
-		}
-	}()
-
-	level.Info(logger).Log("msg", "start index cache processing")
-
-	for _, meta := range metas {
-		// New version of compactor pushes index cache along with data block.
-		// Skip uncompacted blocks.
-		if meta.Compaction.Level == 1 {
-			continue
-		}
-
-		if err := generateIndexCacheFile(ctx, bkt, logger, dir, meta); err != nil {
-			return err
-		}
-		genIndex.Inc()
-	}
-
-	level.Info(logger).Log("msg", "generating index cache files is done, you can remove startup argument `index.generate-missing-cache-file`")
-	return nil
-}
-
-func generateIndexCacheFile(
-	ctx context.Context,
-	bkt objstore.Bucket,
-	logger log.Logger,
-	indexCacheDir string,
-	meta *metadata.Meta,
-) error {
-	id := meta.ULID
-
-	bdir := filepath.Join(indexCacheDir, id.String())
-	if err := os.MkdirAll(bdir, 0777); err != nil {
-		return errors.Wrap(err, "create block dir")
-	}
-
-	defer func() {
-		if err := os.RemoveAll(bdir); err != nil {
-			level.Error(logger).Log("msg", "failed to remove index cache directory", "path", bdir, "err", err)
-		}
-	}()
-
-	cachePath := filepath.Join(bdir, block.IndexCacheFilename)
-	cache := path.Join(meta.ULID.String(), block.IndexCacheFilename)
-
-	ok, err := bkt.Exists(ctx, cache)
-	if ok {
-		return nil
-	}
-	if err != nil {
-		return errors.Wrapf(err, "attempt to check if a cached index file exists")
-	}
-
-	level.Debug(logger).Log("msg", "make index cache", "block", id)
-
-	// Try to download index file from obj store.
-	indexPath := filepath.Join(bdir, block.IndexFilename)
-	index := path.Join(id.String(), block.IndexFilename)
-
-	if err := objstore.DownloadFile(ctx, logger, bkt, index, indexPath); err != nil {
-		return errors.Wrap(err, "download index file")
-	}
-
-	if err := indexheader.WriteJSON(logger, indexPath, cachePath); err != nil {
-		return errors.Wrap(err, "write index cache")
-	}
-
-	if err := objstore.UploadFile(ctx, logger, bkt, cachePath, cache); err != nil {
-		return errors.Wrap(err, "upload index cache")
-	}
-	return nil
-}
-
 type compactConfig struct {
 	haltOnError                                    bool
 	acceptMalformedIndex                           bool
@@ -540,7 +428,6 @@ type compactConfig struct {
 	retentionRaw, retentionFiveMin, retentionOneHr model.Duration
 	wait                                           bool
 	waitInterval                                   time.Duration
-	generateMissingIndexCacheFiles                 bool
 	disableDownsampling                            bool
 	blockSyncConcurrency                           int
 	compactionConcurrency                          int
@@ -583,9 +470,6 @@ func (cc *compactConfig) registerFlag(cmd *kingpin.CmdClause) *compactConfig {
 		Short('w').BoolVar(&cc.wait)
 	cmd.Flag("wait-interval", "Wait interval between consecutive compaction runs and bucket refreshes. Only works when --wait flag specified.").
 		Default("5m").DurationVar(&cc.waitInterval)
-
-	cmd.Flag("index.generate-missing-cache-file", "DEPRECATED flag. Will be removed in next release. If enabled, on startup compactor runs an on-off job that scans all the blocks to find all blocks with missing index cache file. It generates those if needed and upload.").
-		Hidden().Default("false").BoolVar(&cc.generateMissingIndexCacheFiles)
 
 	cmd.Flag("downsampling.disable", "Disables downsampling. This is not recommended "+
 		"as querying long time ranges without non-downsampled data is not efficient and useful e.g it is not possible to render all samples for a human eye anyway").
